time tuple all int32

+++ netcdftime/netcdftime.py
@@ -311,12 +311,14 @@ def DateFromJulianDay(JD, calendar='standard'):
     # don't attempt microsecond accuracy, just round to nearest
     # second (issue #330)
     second = np.round(np.clip((F - minute / 1440.) * 86400., 0, None))
-    second = second.astype(np.int64)
 
-    # convert year, month, day to int32
+    # convert year, month, day, hour, minute, second to int32
     year = year.astype(np.int32)
     month = month.astype(np.int32)
     day = day.astype(np.int32)
+    hour = hour.astype(np.int32)
+    minute = minute.astype(np.int32)
+    second = second.astype(np.int32)
 
     # check if input was scalar and change return accordingly
     isscalar = False
